test: Add test for query search score
